JDRAGClothing/requirements.txt
python-dotenv
langchain-chroma
langchain-core
langchain-community
langchain-text-splitters
dashscope
chromadb
streamlit
sqlalchemy